+++ src/config/identity.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | IDEntity Configuration.
+    |--------------------------------------------------------------------------
+    |
+    | Here you may specify your types providers map.
+    |
+    | Format:
+    | <code>
+    |   %some_type_id% => %type_provider_class_name%
+    | </code>
+    |
+    */
+    'extended_types_map' => [
+        // 'ID_TYPE' => \Your\IDEntityType::class
+    ],
+
+];